still ongoing to implement async logger for each worker threads

+++ SGDEngine/Engine/H1EnginePrivate.h
@@ -4,8 +4,8 @@
 typedef unsigned char byte;
 typedef int int32;
 typedef unsigned int uint32;
-typedef long int int64;
-typedef unsigned long int uint64;
+typedef long long int64;
+typedef unsigned long long uint64;
 
 #if _WIN32
 #define SGD_WINDOWS_PLATFORM 1
